core: refine Application interface and add Governance interface (#24)
Add a new Governance interface for interaction with the governance contract.

Also remove the ValidateBlock call in application interface as the application should validate it before putting it into the consensus module.

A new BlockConverter interface is also added. The consensus module should accept the BlockConverter interface in future implementation, and use the Block() function to get the underlying block info.