core: find block in db (#174)
* Make sure block pool is large enough

It's safe to use a larger blockPool when
the number of chains is smaller.

* Construct latticeData via config.
* Seek acked blocks in blockdb when
   unable to find them in memory cache.

In previous implementation, we assume
our cache in memory is enough to perform
DAG's sanity check. However, it's no longer
true when the number of chains might be
changed between rounds.

* Simplify purge.

Remove the relation to purge block by chainStatus.